chore(origin-backend): do not register entities on a separate connection
